Fixed integer casts in JSON functions. (Was truncating longs.)

+++ util/JSON.cpp
@@ -268,7 +268,7 @@ void
 FromJSON( const string & json, short * pI )
 {
     long int i = strtol( json.c_str(), 0, 0 );
-    *pI = static_cast< int8_t >( i );
+    *pI = static_cast< short >( i );
 }
 
 //.............................................................................
@@ -277,7 +277,7 @@ void
 FromJSON( const string & json, unsigned short * pI )
 {
     unsigned long int i = strtoul( json.c_str(), 0, 0 );
-    *pI = static_cast< uint8_t >( i );
+    *pI = static_cast< unsigned short >( i );
 }
 
 //.............................................................................
@@ -286,7 +286,7 @@ void
 FromJSON( const string & json, int * pI )
 {
     long int i = strtol( json.c_str(), 0, 0 );
-    *pI = static_cast< int16_t >( i );
+    *pI = static_cast< int >( i );
 }
 
 //.............................................................................
@@ -295,7 +295,7 @@ void
 FromJSON( const string & json, unsigned int * pI )
 {
     unsigned long int i = strtoul( json.c_str(), 0, 0 );
-    *pI = static_cast< uint16_t >( i );
+    *pI = static_cast< int >( i );
 }
 
 //.............................................................................
@@ -304,7 +304,7 @@ void
 FromJSON( const string & json, long * pI )
 {
     long int i = strtol( json.c_str(), 0, 0 );
-    *pI = static_cast< int32_t >( i );
+    *pI = i;
 }
 
 //.............................................................................
@@ -313,7 +313,7 @@ void
 FromJSON( const string & json, unsigned long * pI )
 {
     unsigned long int i = strtoul( json.c_str(), 0, 0 );
-    *pI = static_cast< uint32_t >( i );
+    *pI = i;
 }
 
 //.............................................................................
@@ -322,7 +322,7 @@ void
 FromJSON( const string & json, long long * pI )
 {
     long long int i = strtoll( json.c_str(), 0, 0 );
-    *pI = static_cast< int64_t >( i );
+    *pI = i;
 }
 
 //.............................................................................
@@ -331,7 +331,7 @@ void
 FromJSON( const string & json, unsigned long long * pI )
 {
     unsigned long long int i = strtoull( json.c_str(), 0, 0 );
-    *pI = static_cast< uint64_t >( i );
+    *pI = i;
 }
 
 //.............................................................................
